phase2/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
sqlalchemy
psycopg2-binary
alembic
python-dotenv
httpx
pytest
python-multipart
python-jose[cryptography]
passlib[bcrypt]
